Isolated inline functions into separate file
