+++ fbpic/particles/elementary_process/compton/numba_methods.py
@@ -6,92 +6,15 @@ This file is part of the Fourier-Bessel Particle-In-Cell code (FB-PIC)
 It defines numba methods that are used in Compton scattering.
 """
 import numba
-from numba import int64
 import math, random
-from scipy.constants import c, m_e, physical_constants
 from fbpic.threading_utils import njit_parallel, prange
-
-# Get additional useful constants
-r_e = physical_constants['classical electron radius'][0]
-PI_RE_2 = math.pi * r_e**2
-INV_MC = 1./( m_e*c )
-
-@numba.njit
-def lorentz_transform( p_in, px_in, py_in, pz_in, gamma, beta, nx, ny, nz ):
-    """
-    Perform a Lorentz transform of the 4-momentum (p_in, px_in, py_in, pz_in)
-    and return the results
-
-    Parameters
-    ----------
-    p_in, px_in, py_in, pz_in: floats
-        The coordinates of the 4-momentum
-    gamma, beta: floats
-        Lorentz factor and corresponding beta of the Lorentz transform
-    nx, ny, nz: floats
-        Coordinates of *normalized* vector that indicates
-        the direction of the transform
-    """
-    p_parallel_in = nx*px_in + ny*py_in + nz*pz_in
-    p_out = gamma * ( p_in - beta * p_parallel_in )
-    p_parallel_out = gamma * ( p_parallel_in - beta * p_in )
-
-    px_out = px_in + nx * ( p_parallel_out - p_parallel_in )
-    py_out = py_in + ny * ( p_parallel_out - p_parallel_in )
-    pz_out = pz_in + nz * ( p_parallel_out - p_parallel_in )
-
-    return( p_out, px_out, py_out, pz_out )
-
-@numba.njit
-def get_scattering_probability(
-    dt, elec_ux, elec_uy, elec_uz, elec_inv_gamma,
-    photon_n, photon_p, photon_beta_x, photon_beta_y, photon_beta_z ):
-    """
-    # TODO:
-    Write transformations
-    """
-    # Get electron intermediate variable
-    elec_gamma = 1./elec_inv_gamma
-
-    # Get photon density and momentum in the rest frame of the electron
-    transform_factor = elec_gamma \
-        - elec_ux*photon_beta_x - elec_uy*photon_beta_y - elec_uz*photon_beta_z
-    photon_n_rest = photon_n * transform_factor
-    photon_p_rest = photon_p * transform_factor
-
-    # Calculate the Klein-Nishina cross-section
-    k = photon_p_rest * INV_MC
-    f1 = 2 * ( 2 + k*(1+k)*(8+k) ) / ( k**2 * (1 + 2*k)**2 )
-    f2 = ( 2 + k*(2-k) ) * math.log( 1 + 2*k ) / k**3
-    sigma = PI_RE_2 * ( f1 - f2 )
-    # Get the electron proper time
-    proper_dt_rest = dt * elec_inv_gamma
-    # Calculate the probability of scattering
-    p = 1 - math.exp( - sigma * photon_n_rest * c * proper_dt_rest )
-
-    return( p )
-
-@numba.njit
-def get_photon_density_gaussian(
-    elec_x, elec_y, elec_z, ct, photon_n_lab_max, inv_laser_waist2,
-    inv_laser_ctau2, laser_initial_z0, gamma_boost, beta_boost ):
-    """
-    # TODO
-    """
-    # Transform electrons coordinates from simulation frame to lab frame
-    elec_zlab = gamma_boost*( elec_z + beta_boost*ct )
-    elec_ctlab = gamma_boost*( ct + beta_boost*elec_z )
-
-    # Get photon density *in the lab frame*
-    photon_n_lab = photon_n_lab_max * math.exp(
-        - 2*inv_laser_waist2*( elec_x**2 + elec_y**2 ) \
-        - 2*inv_laser_ctau2*(elec_zlab - laser_initial_z0 + elec_ctlab)**2 )
-
-    # Get photon density *in the simulation frame*
-    photon_n_sim = gamma_boost*photon_n_lab*( 1 + beta_boost)
-
-    return( photon_n_sim )
-
+# Import the inline functions
+from .inline_functions import lorentz_transform, get_scattering_probability, \
+    get_photon_density_gaussian, INV_MC
+# Compile the inline functions for CPU
+lorentz_transform = numba.njit( lorentz_transform )
+get_scattering_probability = numba.njit( get_scattering_probability )
+get_photon_density_gaussian = numba.njit( get_photon_density_gaussian )
 
 @njit_parallel
 def get_photon_density_gaussian_numba( photon_n, elec_Ntot,
@@ -109,6 +32,7 @@ def get_photon_density_gaussian_numba( photon_n, elec_Ntot,
             laser_initial_z0, gamma_boost, beta_boost )
 
     return( photon_n )
+
 
 @njit_parallel
 def determine_scatterings_numba( N_batch, batch_size, elec_Ntot,
@@ -150,7 +74,7 @@ def determine_scatterings_numba( N_batch, batch_size, elec_Ntot,
                 photon_p, photon_beta_x, photon_beta_y, photon_beta_z )
 
             # Determine the number of photons produced by this electron
-            nscatter = int64(p * ratio_w_electron_photon + random_draw[ip])
+            nscatter = int(p * ratio_w_electron_photon + random_draw[ip])
             # Note: if p is 0, the above formula will return nscatter=0
             # since random_draw is in [0, 1). Similarly, if p is very small,
             # nscatter will be 1 with probabiliy p * ratio_w_electron_photon,
@@ -162,6 +86,7 @@ def determine_scatterings_numba( N_batch, batch_size, elec_Ntot,
             ip = ip + 1
 
     return( nscatter_per_elec, nscatter_per_batch )
+
 
 # Note: This routine is necessarily serial on CPU, since there is
 # no available thread-safe implementation of on-the-fly random generator.
@@ -188,7 +113,7 @@ def scatter_photons_electrons_numba(
         N_max = min( (i_batch+1)*batch_size, elec_Ntot )
         for i_elec in range( i_batch*batch_size, N_max ):
 
-            # Prepare calculation of scattered photons
+            # Prepare calculation of scattered photons from this electron
             if nscatter_per_elec[i_elec] > 0:
 
                 # Prepare recoil momentum (accumulated for the different
@@ -235,7 +160,7 @@ def scatter_photons_electrons_numba(
                     cos_phi = 1.
                     sin_phi = 0.
 
-            # Loop through the number of scattering for this electron
+            # Loop through the number of scatterings for this electron
             for i_scat in range(nscatter_per_elec[i_elec]):
 
                 # Draw scattering angle in the rest frame, from the
@@ -296,10 +221,11 @@ def scatter_photons_electrons_numba(
                 recoil_py += photon_py - new_photon_py
                 recoil_pz += photon_pz - new_photon_pz
 
-                # Create the new photon by copying the electron data
+                # Create the new photon by copying the electron position
                 photon_x[i_photon] = elec_x[i_elec]
                 photon_y[i_photon] = elec_y[i_elec]
                 photon_z[i_photon] = elec_z[i_elec]
+                photon_w[i_photon] = elec_w[i_elec] * inv_ratio_w_elec_photon
                 # The photon's ux, uy, uz corresponds to the actual px, py, pz
                 photon_ux[i_photon] = new_photon_px
                 photon_uy[i_photon] = new_photon_py
@@ -308,7 +234,6 @@ def scatter_photons_electrons_numba(
                 # with the code for the particle pusher and for the
                 # openPMD back-transformed diagnostics)
                 photon_inv_gamma[i_photon] = 1./new_photon_p
-                photon_w[i_photon] = elec_w[i_elec] * inv_ratio_w_elec_photon
 
                 # Update the photon index
                 i_photon += 1
